requirements.txt
tensorflow
flask
numpy
matplotlib
pillow
scipy
